homework1_app/requirements.txt
flask
gunicorn
cryptography
flask-socketio
eventlet==0.30.2
google-generativeai
psycopg2-binary
requests
python-dotenv
beautifulsoup4
Flask-Failsafe